# swagger.yaml
openapi: 3.0.0
info:
  title: Workout Management API
  description: A complete RESTful API backend for a gym workout tracking application built with Node.js, Express, and PostgreSQL.
  version: 1.0.0
  contact:
    email: support@workoutapi.com
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Routine:
      type: object
      properties:
        id:
          type: integer
          example: 1
        id_user:
          type: integer
          example: 1
        name:
          type: string
          example: "PPL Split"
        description:
          type: string
          nullable: true
          example: "Push Pull Legs routine"
        image_url:
          type: string
          nullable: true
          example: "/images/routines/ppl.jpg"
        workouts:
          type: array
          items:
            $ref: '#/components/schemas/RoutineWorkout'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RoutineWorkout:
      type: object
      properties:
        id:
          type: integer
        id_workout:
          type: integer
        number:
          type: integer
          example: 1
        name:
          type: string
          example: "Push Day"
        description:
          type: string
          nullable: true
...
paths:
...
  /api/routines:
    get:
      tags:
        - Routines
      summary: Get all user routines
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Routines retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  routines:
                    type: array
                    items:
                      $ref: '#/components/schemas/Routine'
    post:
      tags:
        - Routines
      summary: Create new routine
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: "PPL Split"
                description:
                  type: string
                  example: "Push Pull Legs routine"
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Routine created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Routine created successfully"
                  routine:
                    $ref: '#/components/schemas/Routine'
  /api/routines/{id}:
    get:
      tags:
        - Routines
      summary: Get routine with workouts
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Routine retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  routine:
                    $ref: '#/components/schemas/Routine'
    put:
      tags:
        - Routines
      summary: Update routine
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Routine updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Routine updated successfully"
                  routine:
                    $ref: '#/components/schemas/Routine'
    delete:
      tags:
        - Routines
      summary: Delete routine
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Routine deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Routine deleted successfully"
  /api/routines/{id}/workouts:
    post:
      tags:
        - Routines
      summary: Add workout to routine
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - workout_id
                - number
              properties:
                workout_id:
                  type: integer
                  example: 1
                number:
                  type: integer
                  example: 1
      responses:
        '201':
          description: Workout added to routine successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Workout added to routine successfully"
                  routine_workout:
                    $ref: '#/components/schemas/RoutineWorkout'
  /api/routines/{routineId}/workouts/{workoutId}:
    delete:
      tags:
        - Routines
      summary: Remove workout from routine
      security:
        - BearerAuth: []
      parameters:
        - name: routineId
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: workoutId
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Workout removed from routine successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Workout removed from routine successfully"
  /api/routines/{routineId}/workouts/{workoutId}/order:
    put:
      tags:
        - Routines
      summary: Update workout order in routine
      security:
        - BearerAuth: []
      parameters:
        - name: routineId
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: workoutId
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - number
              properties:
                number:
                  type: integer
                  example: 2
      responses:
        '200':
          description: Workout order updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Workout order updated successfully"
                  routine_workout:
                    $ref: '#/components/schemas/RoutineWorkout'
...